Always set -auto-approve
Bundles cannot run in an interactive prompt so we need to always run
commands that do not trigger terraform to wait for user input. This
deprecates the autoApprove field on all actions and always sets
-auto-approve flag on terraform commands.

Once people have had a chance to stop setting the autoApprove field in
their porter.yaml, we can update the mixin in a later release to remove
the field entirely.