Improve docker layer reuse with initFile config
* Add initFile configuration setting. This is the file (relative to the
working directory) that should be copied into the bundle before we run
terraform init during porter build. Limiting to a single file instead of
the entire bundle source allows for layer caching when the bundle source
changes. Defaults to copying the entire working directory (terraform),
instead of the entire bundle source,  which is at
least some improvement, but we should encourage people to isolate their
provider declarations into a single file.
* Add workingDir configuration setting. Defaults to terraform and is
relative to the bundle directory. I am not setting it to
/cnab/app/terraform because this allows us to locate the directory on
the host during build.

Fixes #73
Fixes #16

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>