Fix passing non-string variables for other bundle actions
The install action had been updated in #91 to work with complex data types by using a json parser to read and then write out variables when convertig them from porter.yaml values to -var flags.
This ensures that primitive data types are printed without unexpected formatting, such as quotes around a string value, while arrays and objects are printed in json.

This patch applies the same fix to the other bundle actions: upgrade, invoke and uninstall. I've also extracted that common logic into a function with tests so that we can be sure all the actions are using it correctly.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>