Consolidate test bundle and example bundle
The test bundle had a lot of great edge cases and examples for how to use terraform in a bundle that would be useful to make people aware of in our example bundle. I have updated the basic-tf-example bundle with the contents of our test bundle. I also added defaults for all the parameters so that it doesn't require any additional config to run.

This is also good because our example bundle is tested in CI now.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>